Move get_database_path from database.py to directory_manager.py
